+++ client/src/components/util/HelperFunctions.js
@@ -1,0 +1,11 @@
+const isEmptyString = (str) => {
+    if (typeof str === 'string') {
+        return (!str || str.trim().length === 0 );
+    } else {
+        return false;
+    }
+};
+
+module.exports = {
+    isEmptyString
+};