+++ server/conductor-errors.js
@@ -1,0 +1,8 @@
+const conductorErrors = {
+    'err1': "Required field is missing or malformed.",
+    'err2': "Invalid fields supplied.",
+    'err3': "An error occured saving information to the database.",
+    'err4': "This tool is not available to this organization."
+};
+
+module.exports = conductorErrors;